Move ColorHarmonyOptions component files into their own folder
